started refactoring queries interface
